+++ complete/src/main/java/hello/FileController.java
@@ -8,8 +8,6 @@ import java.util.Comparator;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import javax.servlet.ServletContext;
-
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.util.FileCopyUtils;
@@ -18,27 +16,16 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
-import org.springframework.web.context.ServletContextAware;
 
 @Controller
-public class FileUploadController implements ServletContextAware{
+public class FileController {
 	
-	private ServletContext servletContext;
-	
-	//The controller is not abstract, override the ServletContext supertype.
-	@Override
-	public void setServletContext(ServletContext servletContext) {
-		this.servletContext = servletContext;
-	}
-	
+	private servletContext sContext; 
 	
 	@RequestMapping(method = RequestMethod.GET, value = "/upload")
 	public String provideUploadInfo(Model model) {
 		
-		//List the files in the filesystem
-		File rootFolder = new File(servletContext.getRealPath("/"));
-		//List the files in the memory of server
-		//File rootFolder = new File(Application.ROOT);
+		File rootFolder = new File(sContext.getabsoluteDiskPath());
 		List<String> fileNames = Arrays.stream(rootFolder.listFiles())
 			.map(f -> f.getName())
 			.collect(Collectors.toList());
@@ -58,34 +45,18 @@ public class FileUploadController implements ServletContextAware{
 								   @RequestParam("file") MultipartFile file,
 								   RedirectAttributes redirectAttributes) {
 		if (name.contains("/")) {
-			redirectAttributes.addFlashAttribute("message", "Folder separators not allowed");
-			return "redirect:upload";
-		}
-		if (name.contains("/")) {
-			redirectAttributes.addFlashAttribute("message", "Relative pathnames not allowed");
+			redirectAttributes.addFlashAttribute("message", 
+										"Folder separators or Relative pathnames not allowed");
 			return "redirect:upload";
 		}
 		
 		if (!file.isEmpty()) {
 			try {
-				
 				//Store the uploaded file content in filesystem 
-				File filepath = new File(servletContext.getRealPath("/") + "/" + name); //absolute path 
+				File filepath = new File(sContext.getabsoluteDiskPath()+ "/" + name); 
 				file.transferTo(filepath);
 				redirectAttributes.addFlashAttribute("message",
 						"You successfully uploaded " + name + "to the file system!"+"--path:"+filepath.getPath());
-				
-				
-				//Store the uploaded file content in memory on the server 
-				/*
-				BufferedOutputStream stream = new BufferedOutputStream(
-						new FileOutputStream(new File(Application.ROOT + "/" + name)));
-                FileCopyUtils.copy(file.getInputStream(), stream);
-				stream.close();
-				redirectAttributes.addFlashAttribute("message",
-						"You successfully uploaded " + name + "!");
-				*/
-				
 			}
 			catch (Exception e) {
 				redirectAttributes.addFlashAttribute("message",
@@ -101,32 +72,26 @@ public class FileUploadController implements ServletContextAware{
 		return "redirect:upload";
 	}
 	
-	//This method handle the download, modified from the upload one 
+	
 	@RequestMapping(method = RequestMethod.POST, value = "/download")
 	public String handleFileDownload(@RequestParam("name") String name,
 								   @RequestParam("rename") String rename,
 								   RedirectAttributes redirectAttributes) {
 		if ((name.contains("/")) || (rename.contains("/"))) {
-			redirectAttributes.addFlashAttribute("message", "Folder separators not allowed");
-			return "redirect:upload";
-		}
-		if ((name.contains("/")) || (rename.contains("/"))) {
-			redirectAttributes.addFlashAttribute("message", "Relative pathnames not allowed");
+			redirectAttributes.addFlashAttribute("message", 
+										"Folder separators or Relative pathnames not allowed");
 			return "redirect:upload";
 		}
 		
 		if (!name.isEmpty()) {
 			try {
-				//Convert the relative web path to absolute disk path
-				File filepath = new File(servletContext.getRealPath("/"));
+				File filepath = new File(sContext.getabsoluteDiskPath());
 				if(findFile(name,filepath)){
-					String newname = name;
-					//If the user did not give the rename, 
-					//it will rename downloaded file as '...-Copy' to avoid duplicate
+					String newname = name; 
 					if (!rename.isEmpty()){ newname = rename;}
 					else { newname = name +"-Copy";}
-					FileCopyUtils.copy(new File(servletContext.getRealPath("/") + "/" + name),
-									   new File(servletContext.getRealPath("/") + "/" + newname));
+					FileCopyUtils.copy(new File(sContext.getabsoluteDiskPath() + "/" + name),
+									   new File(sContext.getabsoluteDiskPath() + "/" + newname));
 					redirectAttributes.addFlashAttribute("message",
 						"You successfully downloaded " + name + " to the file system!"+"--path:"+filepath.getPath());
 				}
@@ -149,23 +114,4 @@ public class FileUploadController implements ServletContextAware{
 		
 		return "redirect:upload";
 	}
-	
-	//helper method to find the named file in the internal file system
-	public boolean findFile(String name,File file)
-    {
-        File[] list = file.listFiles();
-        if(list!=null) {
-			for (File fil : list) {
-				if (fil.isDirectory()){
-					findFile(name,fil);
-				}
-				else if (name.equalsIgnoreCase(fil.getName())){
-					return true;
-				}
-			}
-		}
-		return false;
-    }
-	
-
 }
